Move network to plural
