+++ src/Submissions/Api/Areas/Admin/Models/Networks/NetworksModel.cs
@@ -1,6 +1,6 @@
 using System.Collections.Generic;
 
-namespace Biobanks.Submissions.Api.Areas.Admin.Models.Network;
+namespace Biobanks.Submissions.Api.Areas.Admin.Models.Networks;
 
 public class NetworksModel
 {
