+++ src/Submissions/Api/Areas/Admin/Views/Networks/Networks.cshtml
@@ -1,5 +1,5 @@
 @using Microsoft.AspNetCore.Http.Extensions
-@model Biobanks.Submissions.Api.Areas.Admin.Models.Network.NetworksModel;
+@model Biobanks.Submissions.Api.Areas.Admin.Models.Networks.NetworksModel;
 
 @{
     ViewBag.Title = "ADAC Admin";
